pyproject.toml
[project]
name = "deadman-switch"
version = "0.1.0"
description = "A deadman switch application with web-based admin and client interfaces"
readme = "README.md"
authors = [
    { name = "Hansel Yapadi", email = "hyapadi@yahoo.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
    "gunicorn>=21.2.0",
    "aiosqlite>=0.21.0",
    "email-validator>=2.2.0",
    "greenlet>=3.2.2",
]

[project.scripts]
deadman-switch = "deadman_switch.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
